quality mod, colors pbr
