+++ src/camera.rs
@@ -1,5 +1,6 @@
 use crate::car::HID;
 use crate::config::Config;
+use crate::quality::far_culling;
 use crate::track::{AsphaltCell, GroundCell};
 use bevy::input::mouse::MouseMotion;
 use bevy::prelude::*;
@@ -34,40 +35,6 @@ impl Plugin for CarCameraPlugin {
             .add_system(grab_mouse)
             .add_system(far_culling)
             .add_system(camera_switch_system);
-    }
-}
-
-#[cfg(any(target_arch = "wasm32", target_os = "ios", target_os = "android"))]
-const VISIBILITY: f32 = 200.;
-#[cfg(not(any(target_arch = "wasm32", target_os = "ios", target_os = "android")))]
-const VISIBILITY: f32 = 750.;
-
-pub fn far_culling(
-    mut pset: ParamSet<(
-        Query<&Transform, With<Camera>>,
-        Query<(&Transform, &mut Visibility, &ComputedVisibility), With<GroundCell>>,
-        Query<(&Transform, &mut Visibility, &ComputedVisibility), With<AsphaltCell>>,
-    )>,
-) {
-    let cam_translation = pset.p0().single().translation;
-
-    for (ground_cell, mut cell_visibility, computed_visibility) in pset.p1().iter_mut() {
-        if (cam_translation - ground_cell.translation).length() > VISIBILITY {
-            if computed_visibility.is_visible_in_view() {
-                *cell_visibility = Visibility::Hidden;
-            }
-        } else {
-            *cell_visibility = Visibility::Inherited;
-        }
-    }
-    for (asphalt_cell, mut cell_visibility, computed_visibility) in pset.p2().iter_mut() {
-        if (cam_translation - asphalt_cell.translation).length() > VISIBILITY {
-            if computed_visibility.is_visible_in_view() {
-                *cell_visibility = Visibility::Hidden;
-            }
-        } else {
-            *cell_visibility = Visibility::Inherited;
-        }
     }
 }
